🛠 Add yb component interfaces 🔧 Moved Student interface to interfaces instead of ts/interfaces
