+++ app/components/characters/FilterDropdown/interfaces/DropdownOption.interface.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+
+export interface DropdownOptionInterface {
+	house: string
+	checked: boolean
+	filters?: string[]
+	className?: string
+	filterHandler: Function
+	setCondition: React.Dispatch<React.SetStateAction<boolean>>
+}
