+++ app/components/characters/Pagination/PaginationItem.tsx
@@ -1,13 +1,11 @@
 import React from 'react'
 import { StyledPaginationItem } from './styles'
+import { PaginationItemInterface } from './interfaces/PaginationItem.interface'
 
 export const PaginationItem = ({
 	pageNumber,
 	currentPage = false,
-}: {
-	pageNumber: number | string
-	currentPage?: boolean
-}) => {
+}: PaginationItemInterface) => {
 	if (currentPage)
 		return (
 			<StyledPaginationItem className='cursor-pointer bg-dark-3 lg:w-12 lg:h-12 w-9 h-9 text-center rounded-xl my-auto lg:py-3 py-1'>
