+++ app/components/characters/Pagination/interfaces/ChangePageButton.interface.tsx
@@ -1,0 +1,3 @@
+export interface ChangePageButtonInterface {
+	className?: string
+}
