+++ app/components/characters/Pagination/interfaces/PaginationItem.interface.tsx
@@ -1,0 +1,4 @@
+export interface PaginationItemInterface {
+	pageNumber: number | string
+	currentPage?: boolean
+}
