+++ app/components/characters/ProfileCard/interfaces/ProfilePictureContainer.interface.tsx
@@ -1,0 +1,3 @@
+export interface ProfilePictureContainerInterface {
+	image: string
+}
