+++ app/components/characters/interfaces/Student.ts
