+++ app/components/characters/interfaces/index.ts
